fix: four production bugs that only Postgres could see
The suite runs on SQLite; production runs on Postgres. Running the suite once
against Postgres failed nine tests across five causes. Two were live
production bugs, one would have been as soon as it was used, and every one
had passed on SQLite.

Staff could not check anyone into a breakout session. checked_in_by and
checked_out_by were declared uuid but hold a users.id, which is a bigint;
Postgres refused the insert, SQLite stores anything. ledger_transactions.
created_by and data_export_requests.requested_by had the same wrong type but
nothing writes them yet. All four are dropped and re-added as integer foreign
keys — Postgres has no uuid-to-bigint cast, and no row could have held a real
value; production had zero rows in all four tables. A schema test now fails on
SQLite if any user-reference column is not an integer.

Outbound webhooks could not be sent. WebhookCall uses BelongsToTenant, which
writes tenant_id on every insert and filters on it on every read, but the
table had no tenant_id column. Its test created its own copy of the table,
with the column, so it tested a schema that did not exist. The column is added
and backfilled from each call's endpoint, and the test now shares the landlord
connection, as a shared-isolation tenant does in production, so it runs on the
real schema and fails without the migration.

The finance screen received its transaction totals as strings: Postgres
returns SUM() as numeric text. Those and two other uncast database sums are
cast.

Recording usage for an anonymous-class job wrote a NUL byte into a JSON
column, which Postgres rejects. It happens in the job middleware's finally
block, so the failure would replace the job's own outcome. The class name is
cut at the NUL.

+++ app/Http/Controllers/Tenant/FinanceController.php
@@ -132,9 +132,10 @@ final class FinanceController extends Controller
         ]);
 
         $stats = [
-            'total_volume' => MerchantTransaction::where('tenant_id', $tenant->id)->where('status', 'succeeded')->sum('amount'),
+            // Cast: Postgres returns SUM() as a numeric string, which JSON sends as text.
+            'total_volume' => (int) MerchantTransaction::where('tenant_id', $tenant->id)->where('status', 'succeeded')->sum('amount'),
             'transaction_count' => MerchantTransaction::where('tenant_id', $tenant->id)->where('type', 'payment')->count(),
-            'refund_volume' => MerchantTransaction::where('tenant_id', $tenant->id)->where('type', 'refund')->sum('amount'),
+            'refund_volume' => (int) MerchantTransaction::where('tenant_id', $tenant->id)->where('type', 'refund')->sum('amount'),
         ];
 
         return Inertia::render('Tenant/Finance/Index', [
